backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-multipart==0.0.6
httpx==0.25.2
celery==5.3.4
redis==5.0.1
hiredis==2.2.3
flower==2.0.1
pywebpush==1.14.0
firebase-admin==6.4.0
cryptography==41.0.7
fastapi-users[sqlalchemy,oauth]==12.1.2
httpx-oauth==1.12.0
asyncpg==0.29.0
aiosqlite==0.19.0
